requirements.txt
mcp[cli]>=1.12.4
openai
python-dotenv